Fixing extensibility blockers in QueryCompiler/ParameterExtractingExpressionVisitor
