+++ src/EFCore/Query/Internal/QueryCompiler.cs
@@ -315,36 +315,6 @@ namespace Microsoft.EntityFrameworkCore.Query.Internal
                         new TransformingExpressionTreeProcessor(ExpressionTransformerRegistry.CreateDefault())
                     })));
 
-        private class EvaluatableExpressionFilter : EvaluatableExpressionFilterBase
-        {
-            private static readonly PropertyInfo _dateTimeNow
-                = typeof(DateTime).GetTypeInfo().GetDeclaredProperty(nameof(DateTime.Now));
-
-            private static readonly PropertyInfo _dateTimeUtcNow
-                = typeof(DateTime).GetTypeInfo().GetDeclaredProperty(nameof(DateTime.UtcNow));
-
-            private static readonly MethodInfo _guidNewGuid
-                = typeof(Guid).GetTypeInfo().GetDeclaredMethod(nameof(Guid.NewGuid));
-
-            private static readonly List<MethodInfo> _randomNext
-                = typeof(Random).GetTypeInfo().GetDeclaredMethods(nameof(Random.Next)).ToList();
-
-            public override bool IsEvaluatableMethodCall(MethodCallExpression methodCallExpression)
-            {
-                if (_guidNewGuid.Equals(methodCallExpression.Method) 
-                    || _randomNext.Contains(methodCallExpression.Method))
-                {
-                    return false;
-                }
-
-                return base.IsEvaluatableMethodCall(methodCallExpression);
-            }
-
-            public override bool IsEvaluatableMember(MemberExpression memberExpression)
-                => !Equals(memberExpression.Member, _dateTimeNow) 
-                    && !Equals(memberExpression.Member, _dateTimeUtcNow);
-        }
-
         private INodeTypeProvider NodeTypeProvider
             => _nodeTypeProvider
                ?? (_nodeTypeProvider = _nodeTypeProviderFactory.Create());
